v5.x-scala-js: move SigmaExceptions descendant classes to sigmastate.exception package
