+++ sc/src/main/scala/org/ergoplatform/validation/ValidationRules.scala
@@ -1,8 +1,9 @@
 package org.ergoplatform.validation
 
 import scalan.util.Extensions.toUByte
-import sigmastate.Values.{ErgoTree, SValue}
+import sigmastate.Values.{SValue, ErgoTree}
 import sigmastate._
+import sigmastate.exceptions.{SerializerException, SigmaException, InterpreterException}
 import sigmastate.lang.exceptions._
 import sigmastate.serialization.OpCodes.OpCode
 import sigmastate.serialization.TypeSerializer.embeddableIdToType
