+++ sc/src/main/scala/sigmastate/eval/Evaluation.scala
@@ -6,10 +6,9 @@ import scalan.RType._
 import sigmastate.SType._
 import sigmastate.Values.SigmaBoolean
 import sigmastate._
-import sigmastate.lang.exceptions.CostLimitException
 import special.Types._
 import debox.cfor
-
+import sigmastate.exceptions.CostLimitException
 import java.math.BigInteger
 import scala.reflect.ClassTag
 import scala.util.Try
