+++ sc/src/main/scala/sigmastate/serialization/DataSerializer.scala
@@ -10,10 +10,10 @@ import sigmastate.Values.SigmaBoolean
 import sigmastate.utils.{SigmaByteReader, SigmaByteWriter}
 import sigmastate._
 import sigmastate.eval.{Evaluation, _}
-import sigmastate.lang.exceptions.SerializerException
 import special.collection._
 import special.sigma._
 import debox.cfor
+import sigmastate.exceptions.SerializerException
 import scala.collection.mutable
 
 /** This works in tandem with ConstantSerializer, if you change one make sure to check the other.*/
