+++ sc/src/main/scala/sigmastate/serialization/ErgoTreeSerializer.scala
@@ -6,13 +6,13 @@ import sigmastate.{SType, VersionContext}
 import sigmastate.Values.{Constant, ErgoTree, UnparsedErgoTree}
 import sigmastate.lang.DeserializationSigmaBuilder
 import sigmastate.lang.Terms.ValueOps
-import sigmastate.lang.exceptions.{SerializerException, ReaderPositionLimitExceeded}
+import sigmastate.lang.exceptions.ReaderPositionLimitExceeded
 import sigmastate.utils.{SigmaByteReader, SigmaByteWriter}
 import sigmastate.Values.ErgoTree.EmptyConstants
 import sigmastate.util.safeNewArray
 import sigmastate.utxo.ComplexityTable
 import debox.cfor
-
+import sigmastate.exceptions.SerializerException
 import java.util
 
 /**
