+++ sc/src/test/scala/sigmastate/serialization/DataJsonEncoderSpecification.scala
@@ -13,7 +13,7 @@ import sigmastate._
 import sigmastate.eval.Extensions._
 import sigmastate.eval.{Evaluation, _}
 import sigmastate.basics.CryptoConstants.EcPointType
-import sigmastate.lang.exceptions.SerializerException
+import sigmastate.exceptions.SerializerException
 import special.sigma.{Box, AvlTree}
 
 class DataJsonEncoderSpecification extends SerializationSpecification {
